requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
click==8.0.1
colorama==0.4.4
dateparser==1.0.0
fastapi==0.67.0
greenlet==1.1.0
h11==0.12.0
itsdangerous==2.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.18.0
python-multipart==0.0.5
pytz==2021.1
regex==2021.7.6
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
tzlocal==2.1
uvicorn==0.14.0
webassets==2.0
Werkzeug==2.0.1